Move sound notification into highlight module
